+++ osidb/tests/endpoints/flaws/test_cvss_scores.py
@@ -7,6 +7,108 @@ from osidb.tests.factories import AffectFactory, FlawCVSSFactory, FlawFactory
 pytestmark = pytest.mark.unit
 
 
+class TestEndpointsFlawsCVSSScoresV2:
+    """
+    Test that editing a FlawCVSS record through the v2 API
+    only works with the correct Issuer (REDHAT).
+    """
+
+    @pytest.mark.enable_signals
+    def test_flawcvss_create(self, auth_client, test_api_v2_uri):
+        flaw = FlawFactory(impact=Impact.LOW)
+        cvss_data = {
+            "cvss_version": FlawCVSS.CVSSVersion.VERSION3,
+            "vector": "CVSS:3.1/AV:L/AC:L/PR:N/UI:R/S:U/C:H/I:H/A:H",
+            "embargoed": flaw.embargoed,
+        }
+
+        response = auth_client().post(
+            f"{test_api_v2_uri}/flaws/{str(flaw.uuid)}/cvss-scores",
+            data=cvss_data,
+            format="json",
+        )
+        assert response.status_code == status.HTTP_201_CREATED
+        # CVSS scores created through API should always be of type Red Hat
+        assert response.data["issuer"] == FlawCVSS.CVSSIssuer.REDHAT
+        assert FlawCVSS.objects.count() == 1
+
+    @pytest.mark.enable_signals
+    def test_flawcvss_rh_update(self, auth_client, test_api_v2_uri):
+        flaw = FlawFactory()
+        cvss = FlawCVSSFactory(
+            flaw=flaw,
+            version=FlawCVSS.CVSSVersion.VERSION3,
+            vector="CVSS:3.1/AV:L/AC:L/PR:N/UI:R/S:U/C:H/I:H/A:N",
+            issuer=FlawCVSS.CVSSIssuer.REDHAT,
+        )
+
+        cvss_data = {
+            "cvss_version": FlawCVSS.CVSSVersion.VERSION3,
+            "vector": "CVSS:3.1/AV:L/AC:L/PR:N/UI:R/S:U/C:H/I:H/A:H",
+            "embargoed": flaw.embargoed,
+            "updated_dt": cvss.updated_dt,
+        }
+
+        response = auth_client().put(
+            f"{test_api_v2_uri}/flaws/{str(flaw.uuid)}/cvss-scores/{cvss.uuid}",
+            data=cvss_data,
+            format="json",
+        )
+        assert response.status_code == status.HTTP_200_OK
+        assert response.data["vector"] == cvss_data["vector"]
+        assert cvss.vector != response.data["vector"]
+
+    @pytest.mark.enable_signals
+    def test_flawcvss_non_rh_update(self, auth_client, test_api_v2_uri):
+        flaw = FlawFactory()
+        cvss = FlawCVSSFactory(
+            flaw=flaw,
+            version=FlawCVSS.CVSSVersion.VERSION3,
+            issuer=FlawCVSS.CVSSIssuer.NIST,
+            vector="CVSS:3.1/AV:L/AC:L/PR:N/UI:R/S:U/C:H/I:H/A:N",
+        )
+
+        cvss_data = {
+            "cvss_version": FlawCVSS.CVSSVersion.VERSION3,
+            "vector": "CVSS:3.1/AV:L/AC:L/PR:N/UI:R/S:U/C:H/I:H/A:H",
+            "embargoed": flaw.embargoed,
+            "updated_dt": cvss.updated_dt,
+        }
+
+        response = auth_client().put(
+            f"{test_api_v2_uri}/flaws/{str(flaw.uuid)}/cvss-scores/{cvss.uuid}",
+            data=cvss_data,
+            format="json",
+        )
+        assert response.status_code == status.HTTP_400_BAD_REQUEST
+        assert "Only Red Hat CVSS scores can be edited" in response.json()["issuer"]
+        refreshed_cvss = FlawCVSS.objects.first()
+        assert refreshed_cvss and refreshed_cvss.vector == cvss.vector
+
+    @pytest.mark.enable_signals
+    def test_flawcvss_rh_delete(self, auth_client, test_api_v2_uri):
+        flaw = FlawFactory()
+        AffectFactory(flaw=flaw)
+        cvss = FlawCVSSFactory(flaw=flaw, issuer=FlawCVSS.CVSSIssuer.REDHAT)
+
+        url = f"{test_api_v2_uri}/flaws/{str(flaw.uuid)}/cvss-scores/{cvss.uuid}"
+        response = auth_client().delete(url, HTTP_BUGZILLA_API_KEY="foo")
+        assert response.status_code == status.HTTP_200_OK
+        assert FlawCVSS.objects.count() == 0
+
+    @pytest.mark.enable_signals
+    def test_flawcvss_non_rh_delete(self, auth_client, test_api_v2_uri):
+        flaw = FlawFactory()
+        AffectFactory(flaw=flaw)
+        cvss = FlawCVSSFactory(flaw=flaw, issuer=FlawCVSS.CVSSIssuer.NIST)
+
+        url = f"{test_api_v2_uri}/flaws/{str(flaw.uuid)}/cvss-scores/{cvss.uuid}"
+        response = auth_client().delete(url, HTTP_BUGZILLA_API_KEY="foo")
+        assert response.status_code == status.HTTP_400_BAD_REQUEST
+        assert "Only Red Hat CVSS scores can be edited" in response.json()["issuer"]
+        assert FlawCVSS.objects.count() == 1
+
+
 class TestEndpointsFlawsCVSSScores:
     """
     tests specific to /flaws/.../cvss_scores endpoint